pyproject.toml
[project]
name = "Cancervision"
version = "0.1.0"
requires-python = "==3.13.*"
dependencies = [
    "monai",
    "torch",
    "torchvision",
    "numpy",
    "scikit-learn",
    "wandb",
    "matplotlib>=3.10.8",
    "onnxruntime>=1.24.3",
    "tdqm>=0.0.1",
    "tqdm>=4.67.3",
    "nibabel>=5.4.0",
    "ignite>=1.1.0",
    "weave>=0.52.30",
    "python-dotenv>=1.1.1",
    "h5py>=3.16.0",
    "flask>=3.1.0",
]

[tool.uv.sources]
torch = [
  { index = "pytorch-cpu", marker = "sys_platform == 'darwin'" },
  { index = "pytorch-cu130", marker = "sys_platform == 'win32'" },
  { index = "pytorch-cu129", marker = "sys_platform == 'linux'" },
]
torchvision = [
  { index = "pytorch-cpu", marker = "sys_platform == 'darwin'" },
  { index = "pytorch-cu130", marker = "sys_platform == 'win32'" },
  { index = "pytorch-cu129", marker = "sys_platform == 'linux'" },
]

[[tool.uv.index]]
name = "pytorch-cu129"
url = "https://download.pytorch.org/whl/cu129"
explicit = true

[[tool.uv.index]]
name = "pytorch-cu130"
url = "https://download.pytorch.org/whl/cu130"
explicit = true

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true
